Add files via upload
I used it our call roster in FMC B/Kebbi

So that any doctor on the call roster can get the information about is calls only, instead of that of everybody's. 

It receives a spreadsheet file and iterate through the cells and check for presence or absence of the searched doctor's and and removes any information not related to hum or her.